% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radish_graph.R
\name{conductance_surface}
\alias{conductance_surface}
\title{Create a parameterized conductance surface}
\usage{
conductance_surface(covariates, coords, directions = 4, saveStack = TRUE)
}
\arguments{
\item{covariates}{A \code{RasterStack} containing spatial covariates}

\item{coords}{A \code{SpatialPoints} object containing coordinates for a set of focal cells, with the same projection as the rasters in \code{covariates}}

\item{directions}{If \code{4}, consider only horizontal/vertical neighbours as adjacent; if \code{8}, also consider diagonal neighbours as adjacent}

\item{saveStack}{If \code{TRUE}, the \code{RasterStack} is returned with missing data masked uniformly across rasters}
}
\value{
An object of class \code{radish_graph}
}
\description{
Given a set of spatial covariates and a set of spatial coordinates, create a
graph representing a parameterized conductance surface.
}
\details{
NAs are shared across rasters in \code{covariates}, and a warning is thrown if a given cell has mixed NA and non-NA values across the stack. Comparing models with different patterns of missing spatial data (e.g. fit to different stacks of rasters) can give superficially inconsistant results, as these essentially involve different sets of vertices. Thus model comparison should use models fitted to the same \code{radish_graph} object.

Disconnected components are identified and removed, so that only the largest connected component in the graph is retained. The function aborts if there are focal cells that belong to a disconnected component.

Rasters of categorical covariates must have an associated RAT to be 'recognized' as categorical, see \code{\link[raster]{ratify}} and \code{examples} below. The names of levels are taken from the \code{VALUE} column of the RAT, if it exists (otherwise, the integer codes in the \code{ID} column are used).
}
\examples{

library(raster)

data(melip)

covariates <- raster::stack(list(altitude=raster::scale(melip.altitude), 
                                 forestcover=raster::scale(melip.forestcover)))

surface <- conductance_surface(covariates, melip.coords, directions = 8)

# categorical covariates:
# rasters of categorical covariates must have an associated RAT, see 'details'
forestcover_class <- cut(raster::values(melip.forestcover), breaks = c(0, 1/3, 1/6, 1)) 
melip.forestcover_cat <- 
  raster::ratify(raster::setValues(melip.forestcover, as.numeric(forestcover_class)))

RAT <- levels(melip.forestcover_cat)[[1]]
RAT$VALUE <- levels(forestcover_class) #explicitly define level names
levels(melip.forestcover_cat) <- RAT

covariates_cat <- raster::stack(list(forestcover = melip.forestcover_cat,
                                     altitude = melip.altitude)) 

}
\references{
Pope NS. In prep. Fast gradient-based optimization of resistance surfaces.
}
\seealso{
\code{\link{radish}}, \code{\link[raster]{stack}}
}
