% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_distance.R
\name{dist_from_biallelic}
\alias{dist_from_biallelic}
\title{Distance from allele counts}
\usage{
dist_from_biallelic(Y, N)
}
\arguments{
\item{Y}{A matrix containing allele counts, of dimension (number of demes) x (number of loci)}

\item{N}{A matrix containing the number of sampled haplotypes, of dimension (number of demes) x (number of loci)}
}
\value{
A matrix containing pairwise distance
}
\description{
Calculates genetic distance from counts of the derived allele across biallelic makers,
using normalized allele frequencies
}
\details{
TODO formula
}
