% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_distance.R
\name{dist_from_cov}
\alias{dist_from_cov}
\title{Distance matrix from covariance matrix}
\usage{
dist_from_cov(Cov)
}
\arguments{
\item{Cov}{A covariance matrix (does not need to be full-rank)}
}
\value{
A distance matrix of the same dimension as the input
}
\description{
Returns the squared-distance matrix associated with a given covariance matrix
}
\details{
TODO formula
}
