% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hager_zhang.R
\name{HagerZhangControl}
\alias{HagerZhangControl}
\title{Control settings for Hager-Zhang line search}
\usage{
HagerZhangControl(
  delta = 0.1,
  sigma = 0.9,
  alphamax = Inf,
  rho = 5,
  epsilon = 1e-06,
  gamma = 0.66,
  linesearchmax = 50,
  psi3 = 0.1,
  c = 1,
  verbose = FALSE
)
}
\description{
TODO after finalizing line search
}
