% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radish_conductance_model.R
\name{linear_conductance}
\alias{linear_conductance}
\title{Identity-link conductance model}
\usage{
linear_conductance(formula, x)
}
\description{
A function of class "conductance_model" that represents a linear
mapping from spatial covariates to conductance
}
\details{
The model is of the form:

  C_i = x_{i1} * theta_{1} + x_{i2} * theta_{2} + ...

where "C_i" is the conductance of vertex "i", "x_ij" is the value of spatial
covariate "j" at vertex "i", and "theta_j" is the parameter associated with
covariate "j". The intercept is omitted as it is non-identifiable.

Categorical covariates are dummy-coded using the default contrasts, via \code{\link[stats]{model.matrix}}.
}
