% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radish_conductance_model.R
\name{loglinear_conductance}
\alias{loglinear_conductance}
\title{Log-link conductance model}
\usage{
loglinear_conductance(x, theta)
}
\description{
A function of class "conductance_model" that represents a log-linear
mapping from spatial covariates to conductance, such that
}
\details{
C_i = exp(x_{i1} * theta_{1} + x_{i2} * theta_{2} + ...)

where "C_i" is the conductance of vertex "i", "x_ij" is the value of spatial
covariate "j" at vertex "i", and "theta_j" is the parameter associated with
covariate "j".
}
