% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpe.R
\name{mlpe}
\alias{mlpe}
\title{Maximum likelihood population effects}
\usage{
mlpe(
  E,
  S,
  phi,
  nu = NULL,
  gradient = TRUE,
  hessian = TRUE,
  partial = TRUE,
  nonnegative = TRUE,
  validate = FALSE
)
}
\arguments{
\item{E}{A submatrix of the generalized inverse of the graph Laplacian (alternatively, a covariance matrix)}

\item{S}{A matrix of observed genetic distances}

\item{phi}{Nuisance parameters (see details)}

\item{nu}{Number of genetic markers (ignored)}

\item{gradient}{Compute gradient of negative loglikelihood wrt phi?}

\item{hessian}{Compute Hessian matrix of negative loglikelihood wrt phi?}

\item{partial}{Compute second partial derivatives of negative loglikelihood wrt phi and spatial covariates/observed genetic distances}

\item{nonnegative}{Force slope to be nonnegative?}

\item{validate}{Numerical validation via 'numDeriv' (very slow, use for debugging small examples)}
}
\value{
A list containing at a minimum:
 \item{covariance}{rows/columns of the generalized inverse of the graph Laplacian for a subset of target vertices}
 \item{objective}{(if 'objective') the negative loglikelihood}
 \item{fitted}{((if 'objective') a matrix of expected genetic distances among target vertices}
 \item{boundary}{(if 'objective') is the MLE on the boundary (e.g. no genetic structure)?}
 \item{gradient}{(if 'gradient') gradient of negative loglikelihood with respect to phi}
 \item{hessian}{(if 'hessian') Hessian matrix of the negative loglikelihood with respect to phi}
 \item{gradient_E}{(if 'partial') gradient with respect to the generalized inverse of the graph Laplacian}
 \item{partial_E}{(if 'partial') Jacobian of 'gradient_E' with respect to phi}
 \item{partial_S}{(if 'partial') Jacobian of 'gradient' with respect to S}
 \item{jacobian_E}{(if 'partial') a function used for reverse algorithmic differentiation}
 \item{jacobian_S}{(if 'partial') a function used for reverse algorithmic differentiation}
}
\description{
A function of class "measurement_model" that calculates likelihood,
gradient, hessian, and partial derivatives of nuisance parameters and the
Laplacian generalized inverse, using the "maximum likelihood population
effects" model of Clarke et al (2002).
}
\details{
The nuisance parameters are the intercept, slope, log residual
standard deviation, and logit correlation parameter of the MLPE regression. If 'phi' is not
supplied, the MLE of phi is returned using package 'corMLPE' (github.com/nspope/corMLPE)

TODO: formula
}
\examples{
library(raster)

data(melip)

covariates <- raster::stack(list(altitude=melip.altitude, forestcover=melip.forestcover))
surface <- radish_conductance_surface(covariates, melip.coords, directions = 8)

# null (IBD) resistance distance
laplacian_inv <- radish_distance(radish::loglinear_conductance, surface, 
                                 theta = matrix(0, 1, 2), covariance = TRUE)$covariance[,,1]

mlpe(laplacian_inv, melip.Fst) #without 'phi': return MLE of phi
mlpe(laplacian_inv, melip.Fst, phi = c(0., 0.5, 0.1, 0.3))

}
\references{
Clarke et al. TODO
}
