% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton_raphson.R
\name{NewtonRaphsonControl}
\alias{NewtonRaphsonControl}
\title{Control settings for Newton-Raphson optimizer}
\usage{
NewtonRaphsonControl(
  maxit = 100,
  ctol = sqrt(.Machine$double.eps),
  ftol = sqrt(.Machine$double.eps),
  etol = 10 * .Machine$double.eps,
  verbose = FALSE,
  eps = 1e-08,
  ls.control = HagerZhangControl()
)
}
\description{
TODO after finalizing line search
}
