% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radish_conductance_model.R
\name{radish_conductance_model_factory}
\alias{radish_conductance_model_factory}
\title{Conductance model factories}
\description{
Functions that generate objects of class "radish_conductance_model", that represent mappings from spatial data (e.g. rasters) to conductance.
}
\seealso{
\code{\link{linear_conductance}}, \code{\link{loglinear_conductance}}
}
