% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radish_graph.R
\name{radish_conductance_surface}
\alias{radish_conductance_surface}
\title{Create a parameterized conductance surface}
\usage{
radish_conductance_surface(
  covariates,
  coords,
  directions = 4,
  saveStack = TRUE
)
}
\arguments{
\item{covariates}{A 'RasterStack' containing spatial covariates}

\item{coords}{A 'SpatialPoints' object containing coordinates for a set of target cells}

\item{directions}{If '4', consider only horizontal/vertical neighbours as adjacent; if '8', also consider diagonal neighbours as adjacent}

\item{saveStack}{If TRUE, the modified RasterStack is returned}
}
\value{
An object of class 'radish_graph'
}
\description{
Given a set of spatial covariates (as a 'RasterStack') and a set of
spatial coordinates (class 'SpatialPoints'), create an abstract parameterized
conductance surface.
}
\details{
NAs are shared across components of the RasterStack, and a warning is thrown if a given cell has mixed NA and non-NA values across the stack. Disconnect components are identified and removed (only the largest connected component is retained).
}
\examples{
library(raster)

data(melip)

covariates <- raster::stack(list(altitude=melip.altitude, forestcover=melip.forestcover))
surface <- radish_conductance_surface(covariates, melip.coords, directions = 8)

}
\references{
Pope NS. In prep. Fast gradient-based optimization of resistance surfaces.
}
