% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radish_server.R
\name{radish_server}
\alias{radish_server}
\title{Radish server}
\usage{
radish_server(input, output, session)
}
\arguments{
\item{input}{ui input}

\item{output}{ui output}

\item{session}{shiny session}
}
\description{
Server of the Radish application
}
