% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_radish_app.R
\name{run_radish_app}
\alias{run_radish_app}
\title{Runs the Radish application
Interactively or not, dependending on the context.}
\usage{
run_radish_app(...)
}
\arguments{
\item{...}{parameters passed to runApp}
}
\description{
Runs the Radish application
Interactively or not, dependending on the context.
}
