% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_share_app.R
\name{run_share_app}
\alias{run_share_app}
\title{Runs the Share application
Interactively or not, dependending on the context.}
\usage{
run_share_app(...)
}
\arguments{
\item{...}{parameters passed to runApp}
}
\description{
Runs the Share application
Interactively or not, dependending on the context.
}
