% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_server.R
\name{share_server}
\alias{share_server}
\title{Share server}
\usage{
share_server(input, output, session)
}
\arguments{
\item{input}{ui input}

\item{output}{ui output}

\item{session}{shiny session}
}
\description{
Server of the Share application
}
