% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radish_graph.R
\name{conductance_surface}
\alias{conductance_surface}
\title{Create a parameterized conductance surface}
\usage{
conductance_surface(covariates, coords, directions = 4, saveStack = TRUE)
}
\arguments{
\item{covariates}{A 'RasterStack' containing spatial covariates}

\item{coords}{A 'SpatialPoints' object containing coordinates for a set of focal cells}

\item{directions}{If '4', consider only horizontal/vertical neighbours as adjacent; if '8', also consider diagonal neighbours as adjacent}

\item{saveStack}{If TRUE, the modified RasterStack is returned}

\item{formula}{A 'formula' object for the spatial data in 'covariates' (see 'Details')}
}
\value{
An object of class 'radish_graph'
}
\description{
Given a set of spatial covariates (as a 'RasterStack') and a set of
spatial coordinates (class 'SpatialPoints'), create an abstract parameterized
conductance surface.
}
\details{
NAs are shared across rasters in 'covariates', and a warning is thrown if a given cell has mixed NA and non-NA values across the stack. Comparing models with different patterns of missing spatial data (e.g. fit to different stacks of rasters) can give superficially inconsistant results, as these essentially involve different sets of vertices. The recommended way to create sets of nested models (with an invariant set of vertices) is via '?downdate'.

Disconnected components are identified and removed, so that only the largest connected component in the graph is retained. The function aborts if there are focal cells that belong to a disconnected component.

The design matrix for the spatial data that is ultimately used in 'radish' (by an object of class 'conductance_model') is created from 'formula' and 'covariates' via 'model.matrix'.
}
\examples{
library(raster)

data(melip)

covariates <- raster::stack(list(altitude=melip.altitude, forestcover=melip.forestcover))
surface <- radish_conductance_surface(~altitude * forestcover, covariates, melip.coords, directions = 8)

}
\references{
Pope NS. In prep. Fast gradient-based optimization of resistance surfaces.
}
