% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radish_optimize.R
\name{radish}
\alias{radish}
\title{Optimize a parameterized conductance surface}
\usage{
radish(
  f,
  g,
  s,
  S,
  nu = NULL,
  theta = rep(0, ncol(s$x)),
  leverage = TRUE,
  nonnegative = TRUE,
  validate = FALSE,
  control = NewtonRaphsonControl(verbose = TRUE, ctol = 1e-06, ftol = 1e-06)
)
}
\arguments{
\item{f}{A function of class 'conductance_model'}

\item{g}{A function of class 'measurement_model'}

\item{s}{An object of class 'radish_graph'}

\item{S}{A matrix of observed genetic distances}

\item{nu}{Number of genetic markers (potentially used by 'g')}

\item{theta}{Starting values for optimization}

\item{leverage}{Compute influence measures and leverage?}

\item{nonnegative}{Force regression-like 'measurement_model' to have nonnegative slope?}

\item{validate}{Numerical validation via 'numDeriv' (very slow, use for debugging small examples)}

\item{control}{A list containing options for the optimization routine (see ?NewtonRaphsonControl for list)}
}
\value{
An object of class 'radish'
}
\description{
Calculates likelihood, gradient, hessian, and partial derivatives of a
parameterized conductance surface, given a function mapping spatial data to
conductance and a function mapping resistance distance (covariance) to
genetic distance.
}
\examples{
library(raster)

data(melip)

covariates <- raster::stack(list(altitude=melip.altitude, forestcover=melip.forestcover))
surface <- radish_conductance_surface(covariates, melip.coords, directions = 8)

fit_nnls <- radish(radish::loglinear_conductance, radish::leastsquares, surface, melip.Fst)
summary(fit_nnls)

}
