% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radish_grid.R
\name{radish_distance}
\alias{radish_distance}
\title{Resistance distances from a parameterized conductance surface}
\usage{
radish_distance(
  theta,
  formula,
  data,
  conductance_model = radish::loglinear_conductance,
  conductance = TRUE,
  covariance = FALSE
)
}
\arguments{
\item{theta}{A matrix of dimension (grid size) x (number of parameters)}

\item{formula}{A formula with covariates used in the creation of 'data' on the rhs}

\item{data}{An object of class 'radish_graph'}

\item{conductance_model}{A function of class 'radish_conductance_model_factory'}

\item{conductance}{If TRUE, edge conductance is the sum of cell conductances; otherwise edge conductance is the inverse of the sum of cell resistances (NOT USED; TODO)}

\item{covariance}{If TRUE, instead of resistance distances return the associated submatrix of the generalized inverse of graph Laplacian}
}
\value{
An object of class 'radish_grid'
}
\description{
Calculates resistance distances associated with a parameterized conductance surface across
a grid of parameter values.
}
\examples{
library(raster)

data(melip)

covariates <- raster::stack(list(altitude=melip.altitude, forestcover=melip.forestcover))
surface <- radish_conductance_surface(covariates, melip.coords, directions = 8)

theta <- as.matrix(expand.grid(x=seq(-6,6,length.out=21), y=seq(-6,6,length.out=21)))
distances <- radish_distance(radish::loglinear_conductance, surface, theta)

}
