% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radish_grid.R
\name{radish_distance}
\alias{radish_distance}
\title{Resistance distances from a parameterized conductance surface}
\usage{
radish_distance(f, s, theta, covariance = FALSE)
}
\arguments{
\item{f}{A function of class 'conductance_model'}

\item{s}{An object of class 'radish_graph'}

\item{theta}{A matrix of dimension (grid size) x (number of parameters)}

\item{covariance}{If TRUE, instead of resistance distances return the associated submatrix of the generalized inverse of graph Laplacian}
}
\value{
An object of class 'radish_grid'
}
\description{
Calculates resistance distances associated with a parameterized conductance surface across
a grid of parameter values.
}
\examples{
library(raster)

data(melip)

covariates <- raster::stack(list(altitude=melip.altitude, forestcover=melip.forestcover))
surface <- radish_conductance_surface(covariates, melip.coords, directions = 8)

theta <- as.matrix(expand.grid(x=seq(-6,6,length.out=21), y=seq(-6,6,length.out=21)))
distances <- radish_distance(radish::loglinear_conductance, surface, theta)

}
