% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radish_grid.R
\name{radish_grid}
\alias{radish_grid}
\title{Evaluate likelihood of a parameterized conductance surface}
\usage{
radish_grid(
  f,
  g,
  s,
  S,
  theta,
  nu = NULL,
  nonnegative = TRUE,
  covariance = FALSE
)
}
\arguments{
\item{f}{A function of class 'conductance_model'}

\item{g}{A function of class 'measurement_model'}

\item{s}{An object of class 'radish_graph'}

\item{S}{A matrix of observed genetic distances}

\item{theta}{A matrix of dimension (grid size) x (number of parameters)}

\item{nu}{Number of genetic markers (potentially used by 'g')}

\item{nonnegative}{Force regression-like 'measurement_model' to have nonnegative slope?}

\item{covariance}{If TRUE, additionally return (a submatrix of) the generalized inverse of graph Laplacian across the grid}
}
\value{
An object of class 'radish_grid'
}
\description{
Calculates profile likelihood of a parameterized conductance surface across
a grid of parameter values.
}
\examples{
library(raster)

data(melip)

covariates <- raster::stack(list(altitude=melip.altitude, forestcover=melip.forestcover))
surface <- radish_conductance_surface(covariates, melip.coords, directions = 8)

theta <- as.matrix(expand.grid(x=seq(-6,6,length.out=21), y=seq(-6,6,length.out=21)))
grid <- radish_grid(radish::loglinear_conductance, radish::mlpe, surface, melip.Fst, theta, covariance=FALSE)

library(ggplot2)
ggplot(data.frame(loglik=grid$loglik, grid$theta)) + 
  geom_tile(aes(x=x,y=y,fill=-loglik)) + theme_bw() +
  geom_contour(aes(x=x,y=y,z=-loglik), color="black") +
  xlab(expression(theta[altitude])) +
  ylab(expression(theta[forestcover]))

}
